requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-environ==0.8.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
kombu==5.2.3
packaging==21.3
pip==22.0.3
prompt-toolkit==3.0.28
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
psycopg2==2.9.3
channels==3.0.4
